backend/requirements.txt
amqp==5.1.1
asgiref==3.3.2
astroid==2.3.3
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.11.0
isort==4.3.21
kombu==5.2.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-bakery==1.7.0
prompt-toolkit==3.0.30
python-decouple==3.3
pytz==2022.2.1
redis==4.3.3
six==1.13.0
sqlparse==0.3.0
Unipath==1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.11.2
